Move hook file to correct folder
